refactor: 패키지 구조 변경
- web 단에서 도메인별로 controller, dto 분리
- PageResponse는 common.response로 분리
