+++ src/main/java/hanium/modic/backend/domain/post/service/PostService.java
@@ -6,8 +6,8 @@ import hanium.modic.backend.domain.post.entity.PostEntity;
 import hanium.modic.backend.domain.post.entity.PostImageEntity;
 import hanium.modic.backend.domain.post.repository.PostEntityRepository;
 import hanium.modic.backend.domain.post.repository.PostImageEntityRepository;
-import hanium.modic.backend.web.dto.GetPostResponse;
-import hanium.modic.backend.web.dto.PageResponse;
+import hanium.modic.backend.web.post.dto.GetPostResponse;
+import hanium.modic.backend.common.response.PageResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
