+++ src/main/java/hanium/modic/backend/web/post/controller/PostController.java
@@ -1,10 +1,10 @@
-package hanium.modic.backend.web.controller;
+package hanium.modic.backend.web.post.controller;
 
 import hanium.modic.backend.common.response.ApiResponse;
 import hanium.modic.backend.domain.post.service.PostService;
-import hanium.modic.backend.web.dto.CreatePostRequest;
-import hanium.modic.backend.web.dto.GetPostResponse;
-import hanium.modic.backend.web.dto.PageResponse;
+import hanium.modic.backend.web.post.dto.CreatePostRequest;
+import hanium.modic.backend.web.post.dto.GetPostResponse;
+import hanium.modic.backend.common.response.PageResponse;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
