+++ src/test/java/hanium/modic/backend/domain/post/service/PostServiceTest.java
@@ -7,8 +7,8 @@ import hanium.modic.backend.domain.post.entity.PostEntity;
 import hanium.modic.backend.domain.post.entity.PostImageEntity;
 import hanium.modic.backend.domain.post.repository.PostEntityRepository;
 import hanium.modic.backend.domain.post.repository.PostImageEntityRepository;
-import hanium.modic.backend.web.dto.GetPostResponse;
-import hanium.modic.backend.web.dto.PageResponse;
+import hanium.modic.backend.web.post.dto.GetPostResponse;
+import hanium.modic.backend.common.response.PageResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
